requirements.txt
django
djangorestframework
psycopg2-binary
drf-yasg
Pillow
celery[redis]
redis
django-filter
